Split off title and tabset helper
